Move codes to run profiler
