+++ src/cli_tools/mod.rs
@@ -1,1 +1,2 @@
 pub mod dtrace;
+pub mod profiler;
