+++ src/flamegraph/mod.rs
@@ -1,19 +1,16 @@
+use crate::cargo::compile;
+use crate::cargo::CargoTarget;
+use crate::cli_tools::dtrace::make_dtrace_command;
+use crate::cli_tools::profiler::run_profiler;
 use anyhow::bail;
 use anyhow::Context;
 use anyhow::Error;
 use std::fs::OpenOptions;
 use std::io::BufWriter;
 use std::io::Cursor;
-#[cfg(unix)]
-use std::os::unix::process::ExitStatusExt;
 use std::path::Path;
-use std::process::ExitStatus;
 use structopt::StructOpt;
 use tempdir::TempDir;
-
-use crate::cargo::compile;
-use crate::cargo::CargoTarget;
-use crate::cli_tools::dtrace::make_dtrace_command;
 
 mod linux;
 mod macos;
@@ -31,21 +28,6 @@ pub struct FlameGraphCommand {
 
     #[structopt(long)]
     release: bool,
-}
-
-#[cfg(unix)]
-fn terminated_by_error(status: ExitStatus) -> bool {
-    status
-        .signal() // the default needs to be true because that's the neutral element for `&&`
-        .map_or(true, |code| {
-            code != signal_hook::SIGINT && code != signal_hook::SIGTERM
-        })
-        && !status.success()
-}
-
-#[cfg(not(unix))]
-fn terminated_by_error(status: ExitStatus) -> bool {
-    !status.success()
 }
 
 impl FlameGraphCommand {
@@ -73,7 +55,7 @@ impl FlameGraphCommand {
             //
             eprintln!("Profiling {}", binary.path.display());
 
-            let mut cmd = if cfg!(target_os = "macos") {
+            let cmd = if cfg!(target_os = "macos") {
                 make_dtrace_command(
                     root,
                     binary,
@@ -87,38 +69,8 @@ impl FlameGraphCommand {
             } else {
                 bail!("cargo profile flamegraph currently supports only `linux` and `macos`")
             };
-            let cmd_str = format!("{:?}", cmd);
 
-            // Handle SIGINT with an empty handler. This has the
-            // implicit effect of allowing the signal to reach the
-            // process under observation while we continue to
-            // generate our flamegraph.  (ctrl+c will send the
-            // SIGINT signal to all processes in the foreground
-            // process group).
-            #[cfg(unix)]
-            let handler = unsafe {
-                signal_hook::register(signal_hook::SIGINT, || {})
-                    .expect("cannot register signal handler")
-            };
-
-            let mut recorder = cmd
-                .spawn()
-                .with_context(|| format!("failed to spawn: {}", cmd_str))?;
-
-            let exit_status = recorder
-                .wait()
-                .with_context(|| format!("failed to wait for child proceess: {}", cmd_str))?;
-
-            #[cfg(unix)]
-            signal_hook::unregister(handler);
-
-            // only stop if perf exited unsuccessfully, but
-            // was not killed by a signal (assuming that the
-            // latter case usually means the user interrupted
-            // it in some way)
-            if terminated_by_error(exit_status) {
-                bail!("the binary file exited with an error: {}", cmd_str);
-            }
+            run_profiler(cmd).context("failed to profile program")?;
 
             let collapsed: Vec<u8> = if cfg!(target_os = "macos") {
                 crate::cli_tools::dtrace::to_collapsed(
